Add WorkdayHolidaySummarization module
