+++ src/datasets/user-workday-holiday-exp.dataset.ts
@@ -2,15 +2,12 @@ import { Dataset } from './types';
 import * as random from 'random';
 import { Measure, Category, MeasureType, Scope, CategoryValue } from '../models/data-cube/types';
 import { generateCube } from '../models/data-cube/generation';
-import { humanizeMeasureName } from '../utils/formatters';
 import { createTimeSeriesQuery } from './queries/time-series.query';
 import { createLineChartMeta } from './metas/line-chart.meta';
-import { createTabbedChartsMeta } from './metas/tabbed-charts.meta';
 import { PreferenceMeta } from '../services/preference/types';
 import { createDefault } from '../utils/preferences';
 import { DAY } from '../utils/timeUnits';
-import { TimeSeriesPoint } from './queries/time-series.query';
-import { Summary } from './summarizations/types';
+import * as WorkdayHolidaySummarization from './summarizations/workday-holiday-summarization';
 
 export interface Config {
   dailyWeightStd: number;
@@ -70,84 +67,7 @@ export function create(config: Config): Dataset {
 
   const dataCube = generateCube(categories, measures, generateCubeConfig);
 
-  const visitCountSummariesQueryFactory = (points: TimeSeriesPoint[]) => {
-    let summariesCache: Summary[] | null = null;
-    return () => {
-      if (summariesCache) {
-        return summariesCache;
-      }
-
-      const tFunc = (bStart, cStart, cEnd, bEnd) => (v) => {
-        if (v < bStart) {
-          return 0.0;
-        } else if (bStart <= v && v < cStart) {
-          return (v - bStart) / (cStart - bStart);
-        } else if (cStart <= v && v < cEnd) {
-          return 1.0;
-        } else if (cEnd <= v && v < bEnd) {
-          return 1.0 - (v - cEnd) / (bEnd - cEnd);
-        } else {
-          return 0.0;
-        }
-      };
-
-      const tFuncL = (bStart, cStart) => tFunc(bStart, cStart, Infinity, Infinity);
-      const tFuncR = (cEnd, bEnd) => tFunc(-Infinity, -Infinity, cEnd, bEnd);
-
-      const uHighTraffic = ({ y }) => tFuncL(120, 150)(y);
-      const uMediumTraffic = ({ y }) => tFunc(50, 75, 125, 150)(y);
-      const uLowTraffic = ({ y }) => tFuncR(50, 75)(y);
-
-      const uMostPercentage = (v) => tFuncL(0.6, 0.7)(v);
-      const uHalfPercentage = (v) => tFunc(0.3, 0.4, 0.6, 0.7)(v);
-      const uFewPercentage = (v) => tFuncR(0.3, 0.4)(v);
-
-      const uHoliday = ({ x }) => x.getDay() === 5 ? 0.5 : +(x.getDay() === 0 || x.getDay() === 6);
-      const uWeekday = pair => 1 - uHoliday(pair);
-
-      const uTraffics = {
-        high: uHighTraffic,
-        medium: uMediumTraffic,
-        low: uLowTraffic,
-      };
-
-      const uPercentages = {
-        most: uMostPercentage,
-        half: uHalfPercentage,
-        few: uFewPercentage,
-      };
-
-      const uDays = {
-        weekday: uWeekday,
-        holiday: uHoliday,
-      };
-
-      const sigmaCountQAB = (fQ, fA, fB): number => {
-        const uA: number[] = points.map(fA);
-        const uB: number[] = points.map(fB);
-        const n = uA.map((ua, i) => ua * uB[i]).reduce((p, v) => p + v, 0);
-        const d = uA.reduce((p, v) => p + v, 0) + 1e-7;
-        const t = fQ(n / d);
-        return t;
-      };
-
-      const summaries: Summary[] = [];
-      for (const [quantifier, uPercentage] of Object.entries(uPercentages)) {
-        for (const [day, uDay] of Object.entries(uDays)) {
-          for (const [traffic, uTraffic] of Object.entries(uTraffics)) {
-            const t = sigmaCountQAB(uPercentage, uDay, uTraffic);
-            summaries.push({
-              text: `${quantifier} of ${day}s are ${traffic} traffic.`,
-              validity: t
-            });
-          }
-        }
-      }
-
-      summariesCache = summaries;
-      return summaries;
-    };
-  };
+  const visitCountSummariesQueryFactory = WorkdayHolidaySummarization.queryFactory;
 
   const lineChartMeta = createLineChartMeta(
     'Visit Count',
